Title: 자주 등장한 top k 숫자, Time: 11ms, Memory: 1MB, Status: Wrong Answer - CodeTree

+++ 240829/자주 등장한 top k 숫자/top-k-frequent-elements.cpp
@@ -6,24 +6,24 @@
 using namespace std;
 
 int main() {
-    int n, k;
+    long long n, k;
     cin >> n >> k;
 
     long long arr[100000] {};
-    unordered_map<long long, int> unMap;
+    unordered_map<long long, long long> unMap;
     for(int i{ 0 }; i < n; i++) {
         cin >> arr[i];
         unMap[arr[i]]++;
     }
-    map<int, vector<long long>, greater<int>> m;
-    for(unordered_map<long long, int>::iterator it = unMap.begin(); it != unMap.end(); it++) {
+    map<long long, vector<long long>, greater<long long>> m;
+    for(unordered_map<long long, long long>::iterator it = unMap.begin(); it != unMap.end(); it++) {
         int value = it->second;
         if(value >= k) {
             m[value].push_back(it->first);
         }
     }
 
-    for(map<int, vector<long long>, greater<int>>::iterator it = m.begin(); it != m.end(); it++) {
+    for(map<long long, vector<long long>, greater<long long>>::iterator it = m.begin(); it != m.end(); it++) {
         sort(it->second.begin(), it->second.end(), greater<>());
         for(long long ele : it->second) {
             cout << ele << " ";
